+++ src/components/Blog/BlogArticlesBoard/BlogArticlesBoard.tsx
@@ -1,22 +1,18 @@
 'use client';
 
-import { a11yProps, type TabProps } from '@/components/Tabs';
-import type { BlogArticleData, StrapiMetaPagination } from '@/types/strapi';
-import type { Breakpoint, Theme } from '@mui/material';
-import { Box, Skeleton, useMediaQuery, useTheme } from '@mui/material';
+import { type TabProps } from '@/components/Tabs';
+import type { BlogArticleData } from '@/types/strapi';
+import type { Theme } from '@mui/material';
+import { Skeleton, useMediaQuery } from '@mui/material';
 import { useCallback, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import type { GetTagsResponse } from 'src/app/lib/getTags';
-import { TrackingCategory } from 'src/const/trackingKeys';
-import { chunkArray } from 'src/utils/chunkArray';
-import { BlogArticleCard } from '../BlogArticleCard';
 import { BlogArticleCardSkeleton } from '../BlogArticleCard/BlogArticleCardSkeleton';
-import { BlogArticlesTab } from '../BlogArticlesCollections/BlogArticlesTab';
-import { Pagination } from '../Pagination/Pagination';
 import {
   BlogArticlesBoardContainer,
   BlogArticlesBoardTitle,
 } from './BlogArticlesBoard.style';
+import { BlogArticlesBoardTab } from './BlogArticlesBoardTab';
 import { BlogArticlesBoardTabs } from './BlogArticlesBoardTabs';
 
 interface BlogArticlesBoardProps {
@@ -118,6 +114,7 @@ export const BlogArticlesBoard = ({ data, tags }: BlogArticlesBoardProps) => {
               key={`blog-article-card-skeleton-${tabId}-${index}`}
             />
           ),
+          2,
         )
       ) : tags.data?.length > 0 ? (
         tags.data?.map((tag, tagIndex: number) => {
@@ -130,7 +127,7 @@ export const BlogArticlesBoard = ({ data, tags }: BlogArticlesBoardProps) => {
             total: tag.attributes.blog_articles.data.length,
           };
           return (
-            <CategoryTabPanel
+            <BlogArticlesBoardTab
               value={tabId || 0}
               index={tagIndex}
               tags={tags}
@@ -147,92 +144,3 @@ export const BlogArticlesBoard = ({ data, tags }: BlogArticlesBoardProps) => {
     </BlogArticlesBoardContainer>
   );
 };
-
-interface CategoryTabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-  tags: GetTagsResponse;
-  pagination: StrapiMetaPagination;
-  ariaLabel: string;
-  data: BlogArticleData[];
-}
-
-function CategoryTabPanel({
-  children,
-  value,
-  index,
-  tags,
-  data,
-  pagination,
-}: CategoryTabPanelProps) {
-  const theme = useTheme();
-  const [pageTab, setPageTab] = useState(pagination.page);
-  const chunkedPages = chunkArray(data, pagination.pageSize);
-  return (
-    <Box
-      role="tabpanel"
-      hidden={value !== index}
-      {...a11yProps(ariaLabel, value)}
-    >
-      {/* <BlogCarouselContainer
-        sx={{
-          minWidth: '100%',
-          marginLeft: { sx: 0, md: 0, lg: 0, xl: 0 },
-          [theme.breakpoints.up('md' as Breakpoint)]: {
-            marginTop: theme.spacing(6),
-          },
-          [theme.breakpoints.up('xl' as Breakpoint)]: {
-            marginTop: `${theme.spacing(6)}`,
-          },
-        }}
-      > */}
-      <Box
-        sx={{
-          display: 'flex',
-          justifyContent: 'center',
-          flexDirection: 'column',
-        }}
-      >
-        {chunkedPages.map((page, pageIndex) => (
-          <BlogArticlesTab pageTab={pageTab} index={pageIndex}>
-            {page.map((article, articleIndex: number) => (
-              <BlogArticleCard
-                styles={{
-                  display: 'inline-block',
-                  [theme.breakpoints.up('lg' as Breakpoint)]: {
-                    width: 384,
-                  },
-                }}
-                baseUrl={tags.url}
-                id={article.id}
-                key={`blog-articles-board}-${index}-${articleIndex}`}
-                image={article.attributes.Image}
-                title={article.attributes.Title}
-                slug={article.attributes.Slug}
-                trackingCategory={TrackingCategory.BlogArticlesBoard}
-                content={article.attributes.Content}
-                publishedAt={article.attributes.publishedAt}
-                createdAt={article.attributes.createdAt}
-                tags={article.attributes.tags}
-              />
-            ))}
-          </BlogArticlesTab>
-        ))}
-      </Box>
-      {
-        /* todo: enable pagination*/
-        pagination.pageCount > 0 && (
-          <Pagination
-            isEmpty={pagination.pageCount <= 1}
-            page={pageTab}
-            setPage={setPageTab}
-            pagination={pagination}
-            categoryId={value}
-          />
-        )
-      }
-      {/* </BlogCarouselContainer> */}
-    </Box>
-  );
-}
